DROTH-2702 change numeric to bigint

+++ digiroad2-oracle/src/main/resources/db/migration/V1__create_tables.sql
@@ -281,7 +281,7 @@ CREATE TABLE lane
 CREATE TABLE lane_attribute
 (
     id            bigint NOT NULL,
-    lane_id       numeric(38),
+    lane_id       bigint,
     name          varchar(128),
     value         varchar(128),
     required      char(1) DEFAULT '0',
